feat(configuration-view): Separated responsibilities of use-value-suggestion. Added new helper.
